sbomnix: split deriver lookup from fallback loading
Move deriver lookup helpers into sbomnix.derivers and rename the old
store-path loader to fallback_store.FallbackStore. This keeps the legacy
fallback behavior available while making it explicit that normal SBOM
generation should prefer structured closure data.

Update nixgraph and tests to import the narrower modules directly.
